+++ plasmapy/particles/__init__.py
@@ -38,7 +38,7 @@ from plasmapy.particles.particle_class import (
     Particle,
 )
 from plasmapy.particles.particle_input import particle_input
-from plasmapy.particles.particle_json_utils import (
+from plasmapy.particles.serialization import (
     json_load_particle,
     json_loads_particle,
 )
